### 2017-09-07 #399 @whedon
[PRE REVIEW]: Web interface for reflectivity fitting
**Submitting author:** @mdoucet (<a href="http://orcid.org/0000-0002-5560-6478">Mathieu Doucet</a>)
**Repository:** <a href="https://github.com/neutrons/web_reflectivity" target ="_blank">https://github.com/neutrons/web_reflectivity</a>
**Version:** v1.0
