### 2017-09-14 b3773bc
Update paper.md
