### 2017-10-05 #399 @arfon
@jochym, @bjmorgan, would either of you be willing to review this submission for JOSS?
