### 2017-10-10 68729f4
Remove deletion that causes problem.
