+++ web_reflectivity/fitting/view_util.py
@@ -527,8 +527,6 @@ def evaluate_simultaneous_fit(request, instrument, data_id, run_info):
 
     # Update the remote job info
     simul_fit, _ = SimultaneousFit.objects.get_or_create(user=request.user, fit_problem=fit_problem)
-    if simul_fit.remote_job is not None:
-        simul_fit.remote_job.delete()
     simul_fit.remote_job = job
     simul_fit.save()
 
