### 2017-10-11 56557ff
Add absorption parameter
