+++ README.md
@@ -1,4 +1,4 @@
-[![TRAVISCI](https://travis-ci.org/neutrons/web_reflectivity.svg)](https://travis-ci.org/neutrons/web_reflectivity)
+[![TRAVISCI](https://travis-ci.org/neutrons/web_reflectivity.svg?branch=master)](https://travis-ci.org/neutrons/web_reflectivity)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.596150.svg)](https://doi.org/10.5281/zenodo.596150)
 [![codecov](https://codecov.io/gh/neutrons/web_reflectivity/branch/master/graph/badge.svg)](https://codecov.io/gh/neutrons/web_reflectivity)
 [![readthedocs](https://readthedocs.org/projects/web-reflectivity/badge/?version=latest)](https://web-reflectivity.readthedocs.io/en/latest/?badge=latest)
