### 2017-10-11 5c8a116
Fix tests following model update
