### 2017-10-11 40939b5
Update simultaneous fits for absorption
