### 2017-11-22 #399 @nicoguaro
@arfon, @mdoucet, I think I can review this one. But at this point of year it will probably take longer.
