### 2017-11-22 #399 @arfon
@nicoguaro @mdoucet - see you over in #469 for the actual paper review.
