### 2017-11-22 #469 @nicoguaro
@mdoucet, should I install REFL1D to run a local server?
