### 2017-11-22 #469 @mdoucet
@nicoguaro, it's probably best. The interface will still work without it, but the fits won't run. I usually create myself an environment using the following description:

[https://github.com/neutrons/web_reflectivity/blob/master/test/environment/webrefl_env.yml](https://github.com/neutrons/web_reflectivity/blob/master/test/environment/webrefl_env.yml)

Here's a description of how to run a local server:
[https://github.com/neutrons/web_reflectivity/blob/master/test/README.md](https://github.com/neutrons/web_reflectivity/blob/master/test/README.md)

Thanks for taking time to review this!
