### 2017-12-18 #469 @nicoguaro
I have not worked much on this. Nevertheless, I failed to install it.

Building the dependencies I get

    Failed building wheel for python-ldap
