### 2017-12-18 b79afa4
Make sure test env matches the yml file.
