### 2017-12-18 2e91b4c
Adjust requirements following django 2.0 release
