### 2017-12-19 #469 @mdoucet
I suggest following the test installation instructions here:

[https://github.com/neutrons/web_reflectivity/blob/master/test/README.md](https://github.com/neutrons/web_reflectivity/blob/master/test/README.md)

I've added information to those instructions, so start from the top and let me know which item fails, and which OS you are running on.
