### 2018-01-16 #469 @arfon
@nicoguaro - many thanks for all of your help with this submission. In a private communication the authors have decided to withdraw this submission from JOSS.
